+++ pkg/klocust/constant_test.go
