+++ pkg/klocust/template.go
@@ -26,8 +26,8 @@ import (
 	"github.com/Masterminds/sprig"
 	"gopkg.in/yaml.v2"
 
-	"github.com/DevopsArtFactory/klocust/internal/schemas"
-	"github.com/DevopsArtFactory/klocust/internal/util"
+	"github.com/DevopsArtFactory/klocust/pkg/schemas"
+	"github.com/DevopsArtFactory/klocust/pkg/util"
 )
 
 func renderValuesFile(valuesTemplatePath string, valuesFilePath string, value schemas.LocustValues) (string, error) {
