+++ pkg/kube/handler/delete.go
